Fix issue with range computation
Now we correctly check that each regmap is compliante with the given range.
However, the range reported at the end is the real range (and not the max between the given one and the real one).
This introduce a slight change from the user point of view when only one toml file is used.

Fix issue in the samples configuration. Now range is properly checked and thus report the error present in sample files.